fix autocomplete for launch-ide
this is a kludge fix

the problem is that I created a _UseGetOpt for each autocompletion.
in my mind they are all local functions and the completion will use the right one.
in bash's mind, they are all global and rewriting the last one.
all completions will use the last defined _UseGetOpt function.

A better solution would be to figure out how to make _UseGetOpt a local.
The current kludge solution is to rename each function.